Fixed a warning.
git-svn-id: https://svn.opendtect.org/od/trunk@38553 a985855d-d384-4b45-ae33-31a410998c3c

+++ include/Seis/seisdatapack.h
@@ -84,7 +84,7 @@ public:
 						    const BinDataDesc* bdd=0);
 
     int				nrTrcs() const
-				{ return sampling_.hsamp_.totalNr(); }
+				{ return (int)sampling_.hsamp_.totalNr(); }
     TrcKey			getTrcKey(int globaltrcidx) const;
     int				getGlobalIdx( const TrcKey& tk ) const
 				{ return (int)sampling_.hsamp_.globalIdx(tk); }
